Update Project 3 instruction: vending machine.md
Assignment Objectives 
This assignment will allow you to apply iteration in your code and will allow you to become more experienced 
in using integers, floats, strings, and conditional statements. 
 
Your program will simulate a simple change-maker for a vending machine. 
 
The vending machine will start with a stock of coins and dollars. Then it will repeatedly request the user 
purchase an item (enter the price for an item to be purchased) or to quit.  
 
• If the user chooses to enter a price, the vending machine will accept payment in nickels, dimes, 
quarters, one-dollar and five-dollar bills deposited one at a time.  
o When the user has deposited enough to cover the cost of the item, the program will calculate 
the coins to be dispensed in change. 
 
o Alternately, the user can cancel the purchase up until full payment has been deposited. In that 
case, your program will calculate the coins to be dispensed to refund any partial payment 
already deposited. 
 
o With each purchase, the program will update the stock of coins and dollars. 
 
• If the user chooses to quit. Our vending machine will output the remaining stock of coins and dollars. 
 
More detailed instructions are provided with the interactive example below. 
All change and refunds must be in coins only and must use the minimum number of coins possible.